+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 1.3.118 Release:
+## New Modern Content:
+- Circle Casting (Enable via the settings and use via the generic action)
+- Deck of Illusions
 ## Bug Fixes:
 - Topple
+- Exploding Heals
 
 # 1.3.117 Release:
 ## Bug Fixes: